backend/services/processing-service/requirements.txt
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.6.1
pypdf==4.0.1
asyncio==3.4.3 